requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
SQLAlchemy==2.0.48
Werkzeug==3.0.3
python-dotenv==1.0.1
groq==1.1.1
reportlab==4.2.5
psycopg2-binary==2.9.9
gunicorn==22.0.0
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.3.1
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.2.1
PyMuPDF==1.24.3
pypdfium2==5.6.0
pytesseract==0.3.10
httpx==0.28.1
alembic==1.13.0
MarkupSafe==3.0.3
